2026/01/20/replayable-agent-workflows/requirements.txt
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation==0.42b0
pytest==7.4.3
pytest-cov==4.1.0
